double -> float everywhere
TEST THIS...

+++ interface/pid.h
@@ -1,7 +1,7 @@
 #ifndef PID_H
 #define PID_H
 
-double calculate_PID_stabilize(double pid_roll_setpoint, double measurement, double roll_rate);
-double calculate_PID_rate(double pid_roll_setpoint, double measurement);
+float calculate_PID_stabilize(float pid_roll_setpoint, float measurement, float roll_rate);
+float calculate_PID_rate(float pid_roll_setpoint, float measurement);
 
 #endif // PID_H
